refactor: migrate tenant layout and analytics schedule to server components/actions
