+++ lib/actions/analyticsScheduleActions.ts
@@ -1,0 +1,147 @@
+'use server';
+
+import { auth } from '@clerk/nextjs/server';
+import type { AnalyticsActionResult } from '@/types/actions';
+
+export interface ScheduleReportData {
+  orgId: string;
+  name: string;
+  description?: string;
+  frequency: 'daily' | 'weekly' | 'monthly';
+  recipients: string[];
+  filters: any;
+  metrics: string[];
+}
+
+function calculateNextSendDate(frequency: string): Date {
+  const now = new Date();
+  switch (frequency) {
+    case 'daily':
+      return new Date(now.getTime() + 24 * 60 * 60 * 1000);
+    case 'weekly':
+      return new Date(now.getTime() + 7 * 24 * 60 * 60 * 1000);
+    case 'monthly':
+      const nextMonth = new Date(now);
+      nextMonth.setMonth(nextMonth.getMonth() + 1);
+      return nextMonth;
+    default:
+      return new Date(now.getTime() + 24 * 60 * 60 * 1000);
+  }
+}
+
+export async function scheduleAnalyticsReportAction(
+  data: ScheduleReportData
+): Promise<
+  AnalyticsActionResult<{ reportId: string; nextSendDate: Date }>
+> {
+  try {
+    const { userId } = await auth();
+    if (!userId) {
+      return { success: false, error: 'Unauthorized' };
+    }
+
+    const { orgId, name, description, frequency, recipients, filters, metrics } =
+      data;
+    if (!orgId) {
+      return { success: false, error: 'Organization ID is required' };
+    }
+
+    const nextSendDate = calculateNextSendDate(frequency);
+
+    const scheduledReport = {
+      id: `report_${Date.now()}`,
+      organizationId: orgId,
+      userId,
+      name,
+      description,
+      frequency,
+      recipients,
+      filters,
+      metrics,
+      nextSendDate,
+      isActive: true,
+      createdAt: new Date(),
+    };
+
+    // TODO: Persist scheduledReport to database
+    console.log('Scheduled report created:', scheduledReport);
+
+    return { success: true, data: { reportId: scheduledReport.id, nextSendDate } };
+  } catch (error) {
+    console.error('Schedule report error:', error);
+    return { success: false, error: 'Failed to schedule report' };
+  }
+}
+
+export async function listScheduledReportsAction(
+  orgId: string
+): Promise<
+  AnalyticsActionResult<
+    Array<{
+      id: string;
+      name: string;
+      description?: string;
+      frequency: string;
+      recipients: string[];
+      nextSend: Date;
+      isActive: boolean;
+      lastSent: Date;
+    }>
+  >
+> {
+  try {
+    const { userId } = await auth();
+    if (!userId) {
+      return { success: false, error: 'Unauthorized' };
+    }
+
+    // TODO: Replace with database query
+    const scheduledReports = [
+      {
+        id: 'report_1',
+        name: 'Weekly Performance Report',
+        description: 'Weekly analytics summary for management',
+        frequency: 'weekly',
+        recipients: ['manager@company.com'],
+        nextSend: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000),
+        isActive: true,
+        lastSent: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000),
+      },
+      {
+        id: 'report_2',
+        name: 'Monthly Financial Report',
+        description: 'Monthly revenue and cost analysis',
+        frequency: 'monthly',
+        recipients: ['finance@company.com', 'ceo@company.com'],
+        nextSend: new Date(Date.now() + 30 * 24 * 60 * 60 * 1000),
+        isActive: true,
+        lastSent: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
+      },
+    ];
+
+    return { success: true, data: scheduledReports };
+  } catch (error) {
+    console.error('Get scheduled reports error:', error);
+    return { success: false, error: 'Failed to fetch scheduled reports' };
+  }
+}
+
+export async function deleteScheduledReportAction(
+  orgId: string,
+  reportId: string
+): Promise<AnalyticsActionResult<{}>> {
+  try {
+    const { userId } = await auth();
+    if (!userId) {
+      return { success: false, error: 'Unauthorized' };
+    }
+
+    // TODO: Replace with database delete
+    console.log('Deleting scheduled report:', reportId);
+
+    return { success: true };
+  } catch (error) {
+    console.error('Delete scheduled report error:', error);
+    return { success: false, error: 'Failed to delete scheduled report' };
+  }
+}
